requirements.txt
nba_api==1.4.1
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
boto3==1.35.5
snowflake-connector-python==3.12.0
dbt-snowflake==1.8.2
apache-airflow-providers-snowflake
apache-airflow-providers-amazon